fix: large amounts of stacks slow down synthesis
A large amount of stacks leads to a large body of metadata into the
`manifest.json` file. The file can grow too large to be written to disk
(>512MB), but even if it doesn't fail outright, serializing and writing
and loading a large JSON file takes a lot of CPU time.

This PR splits the metadata for stacks off into a separate file which is
written separately, moving it out of the hot path of working with the
`manifest.json` file. On a tree of 256 stacks and ~40,000 constructs,
this shaves off 5 seconds off of the total synthesis time.